データセット詳細画面改修 ・タブを無くした ・選択中のバージョンに対してデータ追加するように修正 ・ドロワーをダイアログに変更
